Some tweaks to logic
Removed some bugs which had improvements suggestions overwriting known data. e.g. if it had flat roof insulation in the the improvements list that might overwrite pitched roof. Now it needs to be that it only happens if we dont know the building element - most had this "gate" in the logic but not all

Also accounted for some very, very edge cases where spaces or flags were causing some mis-labelling. Impacted a small minority