Removed an IF
There was an IF to check for NA values in the geography (and if there are any to ignore that geog for calcs). Made initially as a sledgehammer to catch BLOCK_ID NAs, but these are dealt with below.

No NAs in PC etc so all it did was slow the process down as R had to check each spatial area for an NA first...